+++ spec/lib/oauth/base_request_spec.rb
@@ -4,13 +4,13 @@ module Doorkeeper::OAuth
   describe BaseRequest do
     let(:access_token) do
       double :access_token,
-             token:              "some-token",
-             expires_in:         "3600",
-             expires_in_seconds: "300",
-             scopes_string:      "two scopes",
-             refresh_token:      "some-refresh-token",
-             token_type:         "bearer",
-             created_at:         0
+             plaintext_token:         "some-token",
+             expires_in:              "3600",
+             expires_in_seconds:      "300",
+             scopes_string:           "two scopes",
+             plaintext_refresh_token: "some-refresh-token",
+             token_type:              "bearer",
+             created_at:              0
     end
 
     let(:client) { double :client, id: '1' }