+++ spec/models/doorkeeper/application_spec.rb
@@ -2,6 +2,7 @@ require 'spec_helper'
 
 module Doorkeeper
   describe Application do
+    let(:clazz) { Doorkeeper::Application }
     let(:require_owner) { Doorkeeper.configuration.instance_variable_set('@confirm_application_owner', true) }
     let(:unset_require_owner) { Doorkeeper.configuration.instance_variable_set('@confirm_application_owner', false) }
     let(:new_application) { FactoryBot.build(:application) }
@@ -120,6 +121,42 @@ module Doorkeeper
       new_application.save
       new_application.secret = nil
       expect(new_application).not_to be_valid
+    end
+
+    context 'with hashing enabled' do
+      include_context 'with application hashing enabled'
+      let(:app) { FactoryBot.create :application }
+
+      it 'holds a volatile plaintext and BCrypt secret' do
+        expect(app.plaintext_secret).to be_a(String)
+        expect(app.secret).not_to eq(app.plaintext_secret)
+        expect { BCrypt::Password.create(app.secret) }.not_to raise_error
+      end
+
+      it 'does not fallback to plain lookup by default' do
+        lookup = clazz.by_uid_and_secret(app.uid, app.secret)
+        expect(lookup).to eq(nil)
+
+        lookup = clazz.by_uid_and_secret(app.uid, app.plaintext_secret)
+        expect(lookup).to eq(app)
+      end
+
+      context 'with fallback enabled' do
+        include_context 'with token hashing and fallback lookup enabled'
+
+        it 'provides plain and hashed lookup' do
+          lookup = clazz.by_uid_and_secret(app.uid, app.secret)
+          expect(lookup).to eq(app)
+
+          lookup = clazz.by_uid_and_secret(app.uid, app.plaintext_secret)
+          expect(lookup).to eq(app)
+        end
+      end
+
+      it 'does not provide access to secret after loading' do
+        lookup = clazz.by_uid_and_secret(app.uid, app.plaintext_secret)
+        expect(lookup.plaintext_secret).to be_nil
+      end
     end
 
     describe 'destroy related models on cascade' do